Move rendering setup from Oculus VRViewManager to OculusVR, and implement cleanup
